+++ cmd/tempo/app/config_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/grafana/tempo/tempodb/backend/s3"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/grafana/tempo/modules/distributor"
@@ -37,6 +38,9 @@ func TestConfig_CheckConfig(t *testing.T) {
 						Block: &common.BlockConfig{
 							Version: "v2",
 						},
+						S3: &s3.Config{
+							NativeAWSAuthEnabled: true,
+						},
 					},
 				},
 				Distributor: distributor.Config{
@@ -52,6 +56,7 @@ func TestConfig_CheckConfig(t *testing.T) {
 				warnStorageTraceBackendS3,
 				warnBlocklistPollConcurrency,
 				warnLogReceivedTraces,
+				warnNativeAWSAuthEnabled,
 			},
 		},
 		{